requirements.txt
Django>=4.2.0
django-crispy-forms>=2.0
firebase-admin>=6.2.0
google-cloud-aiplatform>=1.36.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.21.0
opencv-python>=4.8.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
